+++ lib/vy_date.dart
@@ -5,3 +5,4 @@ library vy_date;
 
 export 'src/date.dart';
 export 'src/date_period.dart';
+export 'src/date_period_assembler.dart';